requirements.txt
python-dotenv
groq
AppOpener
bs4
pywhatkit
pillow
requests
rich
keyboard
cohere
googlesearch-python
selenium
mtranslate
pygame
edge-tts
PyQt5
webdriver-manager
pycaw
fastapi
uvicorn
sqlalchemy
alembic
bcrypt
python-jose[cryptography]
python-multipart
passlib
